CRUD User Done

+++ application/views/User/user_list.php
@@ -27,39 +27,69 @@
 	          	<?php //endfor; ?>
 	          </tbody>
 	        </table>
+	        <div id="loading_firebase">
+	        	<div class="loader my-4"></div>
+	        	<h5 class="text-center">Loading...</h5>
+	        </div>
 	      </div>
     	</div>
     </div>
   </div>
 </div>
 <script type="text/javascript">
-	db.collection("Users").get().then((querySnapshot) => {
-		var json_all = [];
-    querySnapshot.forEach((doc) => {
-        // console.log(`${doc.id} => ${doc.data()}`);
-        // console.log(doc.id);
-        // console.log(doc.data());
+	load_data();
 
-        var data = doc.data();
-        // console.log(myJSON);
-        // console.log(doc.id);
-        var json_arr = [data.Nama || "", data.Email || "", data.Role || "", data.Departemen || "", data.Password || "", "<button type='button' class='btn btn-sm btn-danger' onclick='delete_data(this)' data-key='"+doc.id+"'><i class='fa fa-trash'></i></button>"];
-        json_all.push(json_arr);
-    });
-    console.log(JSON.stringify(json_all));
-    $('.datatables').DataTable({
-    	responsive: true,
-    	data: json_all,
-    	order:[]
-    })
-	});
+	function load_data() {
+		$('#loading_firebase').show();
+		db.collection("Users").get().then((querySnapshot) => {
+			var json_all = [];
+	    querySnapshot.forEach((doc) => {
+	        // console.log(`${doc.id} => ${doc.data()}`);
+	        // console.log(doc.id);
+	        // console.log(doc.data());
 
-	function delete_data(btn) {
-		var key = $(btn).attr("data-key");
-		db.collection("Users").doc(key).delete().then(function() {
-	    alert("Document successfully deleted!");
+	        var data = doc.data();
+	        // console.log(myJSON);
+	        // console.log(doc.id);
+	        var action = "<button type='button' class='btn btn-sm btn-danger' onclick='delete_data(this)' data-key='"+doc.id+"'><i class='fa fa-trash'></i></button>"+
+	        	"&nbsp;<a href='<?php echo base_url() ?>user/user_edit/"+doc.id+"' class='btn btn-sm btn-warning'><i class='fa fa-pencil'></i></a>";
+	        var json_arr = [data.Nama || "", data.Email || "", data.Role || "", data.Departemen || "", data.Password || "", action];
+	        json_all.push(json_arr);
+	    });
+
+	    console.log(JSON.stringify(json_all));
+	    $('.datatables').DataTable().clear().destroy();
+	    $('.datatables').DataTable({
+	    	responsive: true,
+	    	data: json_all,
+	    	order:[]
+	    }).draw();
+	    $('#loading_firebase').hide();
 		}).catch(function(error) {
-	    alert("Error removing document: ", error);
+		  console.log("Error getting document:", error);
 		});
+	}
+
+	async function delete_data(btn) {
+		Swal.fire({
+      title: 'Are you sure to <b class="text-danger">&nbsp;Delete&nbsp;</b> this?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, Delete it!'
+    }).then((result) => {
+      if (result.value) {
+      	sweetalert('loading', 'Please Wait...');
+        var key = $(btn).attr("data-key");
+				db.collection("Users").doc(key).delete().then(function() {
+					load_data();
+			    sweetalert('success', 'Document successfully deleted!');
+				}).catch(function(error) {
+			    sweetalert('error', error);
+				});
+      }
+    })
 	}
 </script>
